rename: Test Fixture 클래스를 각 도메인의 fixture 패키지로 이동
