+++ src/test/java/es/princip/getp/domain/client/controller/MyClientControllerTest.java
@@ -4,10 +4,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import es.princip.getp.domain.client.domain.entity.Client;
 import es.princip.getp.domain.client.dto.request.UpdateClientRequest;
 import es.princip.getp.domain.client.exception.ClientErrorCode;
+import es.princip.getp.domain.client.fixture.ClientFixture;
 import es.princip.getp.domain.client.service.ClientService;
 import es.princip.getp.domain.member.domain.entity.Member;
 import es.princip.getp.domain.member.domain.enums.MemberType;
-import es.princip.getp.fixture.ClientFixture;
 import es.princip.getp.global.config.SecurityConfig;
 import es.princip.getp.global.config.SecurityTestConfig;
 import es.princip.getp.global.exception.BusinessLogicException;
