+++ src/test/java/es/princip/getp/domain/project/repository/ProjectLikeRepositoryTest.java
@@ -19,11 +19,11 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.test.context.ActiveProfiles;
 
+import static es.princip.getp.domain.client.fixture.ClientFixture.createClient;
 import static es.princip.getp.domain.member.domain.enums.MemberType.ROLE_CLIENT;
 import static es.princip.getp.domain.member.domain.enums.MemberType.ROLE_PEOPLE;
+import static es.princip.getp.domain.member.fixture.MemberFixture.createMember;
 import static es.princip.getp.domain.people.fixture.PeopleFixture.createPeople;
-import static es.princip.getp.fixture.ClientFixture.createClient;
-import static es.princip.getp.fixture.MemberFixture.createMember;
 import static es.princip.getp.fixture.ProjectFixture.createProject;
 import static org.assertj.core.api.Assertions.assertThat;
