+++ src/test/java/es/princip/getp/domain/project/service/ProjectServiceTest.java
@@ -1,18 +1,13 @@
 package es.princip.getp.domain.project.service;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.BDDMockito.given;
-
 import es.princip.getp.domain.client.domain.entity.Client;
+import es.princip.getp.domain.client.fixture.ClientFixture;
 import es.princip.getp.domain.client.service.ClientService;
 import es.princip.getp.domain.member.domain.entity.Member;
-import es.princip.getp.domain.project.dto.request.CreateProjectRequest;
+import es.princip.getp.domain.member.fixture.MemberFixture;
 import es.princip.getp.domain.project.domain.entity.Project;
+import es.princip.getp.domain.project.dto.request.CreateProjectRequest;
 import es.princip.getp.domain.project.repository.ProjectRepository;
-import es.princip.getp.fixture.ClientFixture;
-import es.princip.getp.fixture.MemberFixture;
 import es.princip.getp.fixture.ProjectFixture;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -21,6 +16,11 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.BDDMockito.given;
 
 @ExtendWith(MockitoExtension.class)
 public class ProjectServiceTest {
